moved motif to motif.rs
